+++ packages/storybook/src/utilities/storybook.ts
@@ -1,6 +1,10 @@
 import { html, ViewTemplate } from '@ni/fast-element';
+import { anchorTag } from '@ni/nimble-components/dist/esm/anchor';
 import { themeProviderTag } from '@ni/nimble-components/dist/esm/theme-provider';
-import { bodyFont } from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
+import {
+    bodyFont,
+    warningFontColor
+} from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
 import type { Theme } from '@ni/nimble-components/dist/esm/theme-provider/types';
 import { listOptionTag } from '@ni/nimble-components/dist/esm/list-option';
 import { listOptionGroupTag } from '@ni/nimble-components/dist/esm/list-option-group';
@@ -151,27 +155,27 @@ export interface OkWarningConfig {
 export const incubatingWarning = (config: IncubatingWarningConfig): string => `
 <style class="code-hide">
 #incubating-warning {
-    color: red;
+    color: var(${warningFontColor.cssCustomProperty});
     font: var(${bodyFont.cssCustomProperty});
     padding-bottom: 16px;
 }
 </style>
 <div id="incubating-warning" class="code-hide">
 WARNING - The ${config.componentName} is still incubating. It is not recommended for application use. 
-See the <a href="${config.statusLink}">incubating component status</a>.
+See the <${anchorTag} href="${config.statusLink}">incubating component status</${anchorTag}>.
 </div>`;
 
 export const okWarning = (config: OkWarningConfig): string => `
 <style class="code-hide">
 #ok-warning {
-    color: #c00000;
+    color: var(${warningFontColor.cssCustomProperty});
     font: var(${bodyFont.cssCustomProperty});
     padding-bottom: 16px;
 }
 </style>
 <div id="ok-warning" class="code-hide">
 WARNING - The ${config.componentName} is an ok component. These components have varying levels of quality and stability. Clients should consult with component-specific owners before leveraging in new applications.
-See the <a href="${config.statusLink}">ok component status</a>.
+See the <${anchorTag} href="${config.statusLink}">ok component status</${anchorTag}>.
 </div>`;
 
 // On the Docs page, there is a div with a scale(1) transform that causes the dropdown to be