+++ packages/blazor-workspace/OkBlazor/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @ni/ok-blazor
 
-<!-- This log was last generated on Tue, 28 Apr 2026 14:27:06 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 05 May 2026 14:11:16 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.2.1
+
+Tue, 05 May 2026 14:11:16 GMT
+
+### Patches
+
+- Bump @ni/ok-components to v1.2.1
+
+## 1.2.0
+
+Mon, 04 May 2026 17:04:06 GMT
+
+### Minor changes
+
+- Bump @ni/ok-components to v1.2.0
+
+## 1.1.2
+
+Sat, 02 May 2026 02:01:43 GMT
+
+### Patches
+
+- Dependency update May 2026 ([ni/nimble@ec45dce](https://github.com/ni/nimble/commit/ec45dce8237675f6f19e17e142532070b6551ab2))
+- Bump @ni/ok-components to v1.1.2
+
+## 1.1.1
+
+Thu, 30 Apr 2026 22:40:17 GMT
+
+### Patches
+
+- Bump @ni/ok-components to v1.1.1
 
 ## 1.1.0